requirements.txt
Django==2.1.1
Pillow==5.2.0
beautifulsoup4==4.6.3
common-framework==2018.7.25
dj-database-url==0.5.0
django-compressor==2.2
django-configurations==2.1
django-cors-headers==2.4.0
django-debug-toolbar==1.10
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework-xml==1.3.0
flake8==3.5.0
ipython==6.5.0
lxml==4.2.4
pytest==3.7.4
pytest-cov==2.6.0
pytest-django==3.4.2
python-dateutil==2.7.3
pip==18.0
pytz==2018.5
requests==2.19.1
setuptools==40.2.0
wheel==0.31.1
